aqua-web: simplify pipeline handling logic
pipelines are now a series of plugs which simply mutate a stateful connection object in turn.
plugs can register callbacks to be fired when the pipeline is finished processing the current request.
this is useful to simulate "stack unwinding" / "around middlewares" while keeping the request processing iterative
w.r.t the call stack.

the beauty of the callbacks is that they are just plugs, so they compose the same as any other member
of the pipeline would.