add associaion between entry tags & tags
NOTE: At the moment this is sort of hacky since diesel doesn't
support "has many through"-style associations.

This moves the models into their own modules to prevent them from
clashing with one another. (Not sure if this is actually a problem, but
I was getting some really gnarly error messages from diesel.)

Next we add an ID to EntryTag, though `infer_schema` seems to understand
the composite key, it did not appear that `#[derive(Identifiable)]`
understood our composite key.

Then we setup an association between an entry tag & a tag, this way we
can join the tables together. This provides a quick way for diesel to go
from an Entry => [Tag]