add detection using ffprobe/ffmpeg
this calls out to system ffprobe and ffmpeg to handle mime type detection:

first we use ffprobe to see if the installation knows how to deal with the filetype we've passed it.
the resulting format is mapped to a mime type & extension, and then ffmpeg is used to create the thumbnail.

this should work pretty well, I just need to add mime mappings for the various ffmpeg format strings.