(fix) add migration for is_orphan
This was a change made outside of diesel.

The aqua_csharp WPF GUI project has a utility (aqua_sister_agnes) which walks the entire content store
and then determines which database entries are not currently in the content store.

The diesel model was updated to reflect this change; but a corresponding migration was not added
mostly because I'm lazy.